celes-desktop/requirements.txt
pywebview>=4.4.1
pyinstaller>=6.0.0
celes>=0.1.4